django/requirements.txt
Django>=3.2,<4.0
django-extensions==3.0.9
psycopg2-binary>=2.8
black==20.8b1
django-spa==0.2.0
flake8==3.8.4
factory_boy==3.2.0
requests==2.24.0
djangorestframework==3.12.1
djangorestframework-gis==0.16
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
